Remove TransactionListFragment
