--- library/src/main/res/layout/chucker_fragment_transaction_list.xml
@@ -1,47 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/transactionsRecyclerView"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:scrollbars="vertical"
-        app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-        tools:context="com.chuckerteam.chucker.internal.ui.transaction.TransactionListFragment"
-        tools:listitem="@layout/chucker_list_item_transaction" />
-
-    <LinearLayout
-        android:id="@+id/tutorialView"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_margin="@dimen/chucker_doub_grid"
-        android:gravity="center_vertical"
-        android:orientation="vertical"
-        android:visibility="gone">
-
-        <TextView
-            style="@style/Chucker.TextAppearance.ListItem"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="@string/chucker_setup" />
-
-        <TextView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="@dimen/chucker_base_grid"
-            android:gravity="center_vertical"
-            android:text="@string/chucker_network_tutorial" />
-
-        <TextView
-            android:id="@+id/tutorialLink"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="@dimen/chucker_base_grid"
-            android:gravity="center_vertical"
-            android:text="@string/chucker_check_readme" />
-    </LinearLayout>
-</FrameLayout>
